ragas/pyproject.toml
[project]
name = "ragas"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=3.0.0",
    "langchain-openai>=0.2.0",
    "langfuse>=2.50.2",
    "llama-index>=0.11.11",
    "python-dotenv>=1.0.1",
    "ragas>=0.1.0",
]

[tool.uv.sources]
ragas = { workspace = true }
